TR-1512 Migrate some tests to separate modal component
