+++ src/pages/templatesV2/components/editorActions/tests/DuplicateTemplateModal.test.js
@@ -1,0 +1,46 @@
+import React from 'react';
+import { shallow } from 'enzyme';
+import useEditorContext from '../../../hooks/useEditorContext';
+import DuplicateTemplateModal from '../DuplicateTemplateModal';
+
+jest.mock('../../../hooks/useEditorContext');
+
+describe('DuplicateTemplateModal', () => {
+  const subject = (editorState, props) => {
+    useEditorContext.mockReturnValue({
+      hasDraft: true,
+      draft: {
+        name: 'My Draft',
+        id: 'my-draft'
+      },
+      ...editorState
+    });
+
+    return shallow(<DuplicateTemplateModal {...props}/>);
+  };
+
+  it('has a close button', () => {
+    const wrapper = subject();
+    const modalProps = wrapper.find('Modal').props();
+
+    expect(modalProps.showCloseButton).toBe(true);
+  });
+
+  it('has a `Panel` component with the title "Duplicate Template"', () => {
+    const wrapper = subject();
+
+    expect(wrapper.find('Panel').props().title).toBe('Duplicate Template');
+  });
+
+  it('has a two `TextField` components', () => {
+    const wrapper = subject();
+
+    expect(wrapper.find('TextField')).toHaveLength(2);
+  });
+
+  it('has a "Duplicate" button', () => {
+    const wrapper = subject();
+
+    expect(wrapper.find('Button')).toExist();
+  });
+});
